refactor(sdk): change Executor Parameters and Configs to a Schema
Instead of having an interface for these types, a Schema is now returned
with all of the necessary metadata of the Parameters inside. This
changes how we load flags from and into cobra.Commands: all flags are
added as JSON values, so they're all treated as strings. When
loading the values back, the JSON string is unmarshalled and the raw
data is sent to Executor.Execute.

[#6]