feat(login): improve user communication, allow fail to open the browser
Be more precise about each step, print out the URL so we can
work-around a missing browser (ex: in a remote SSH without a browser).

Then do not fail if we can't open the browser, just say why it failed
and explain the user should do it manually.

Last but not least, log all messages to stderr. Let's keep stdout only
for the actual results.