feat: enable executors to provide a default output formatter
Some actions will likely format themselves in some more friendly way
unless told so to use something else (yaml, json, template...)

This is done by implementing a ExecutorResultFormatter.

This can be done by either creating an Executor with the
DefaultFormatResult() method, or by wrapping it with a
NewExecuteFormat(), which is more convenient.

In the future we may add some extension to the openapi where we can
state some formatting rules that will generate a specific formatter
function and the runtime will do this wrapping for us.

Right now we're showcasing it with a tabular format for the config
list.