feat(sdk/openapi): improve operation name generation
The old code was producing too-long operation names when it wasn't
needed, however it was allowing duplicated names, which was confusing.

Given that we consier a resource = tag, if the same tag had the same
getActionName() result, it would generate duplicates. Say different
paths with simple "GET", they would both generate operations with the
same name "get".

This commit builds a tree where each level is the path component and
the leafs are the operationDesc (http method, key, pathItem and
operation).

We build the whole tree once and then we walk that to build the
operation name from the path. If the parent of a given element
contains only one child, it's omitted. If it contains multiple
children, then it's appended.

This allows the operations to be as short as possible if no
duplicates, or as long as required by using the whole path.

This commit also renames some HTTP methods to more human-friendly names:
 - "get" becomes "list" if doesn't end with path parameter ("{...}")
 - "post" becomes "create"
 - "patch" becomes "update"
 - "put" becomes "replace"