feat(cli): load flags starting with '@' from files
Pretty much like cURL, any flag which value starts with @ will load
its value from the filename following the at-sign.

To use a plain string that starts with @, escape with quotes.

The file is assumed to be JSON, if it's not, then it's handled as a
string (no stripping or the likes)

Closes: #152