fix(cli): save any value to flag and avoid viper bind
When Viper reads the flag from cobra, it will always get the string
representation rather than the actual unmarshalled type. Thus, we are
saving the unmarshalled type (any) to the `AnyFlagValue` and read
directly from it if set via cmd line. When Viper reads from config file,
it will correctly unmarshall the value to the expected type, thus we
use as a fallback if not set by Cobra.