feat(cli): easier handling of logging config
We must initialize the zapConfig with some values before we unmarshal
on top of it, otherwise the user must provide all of the
configurations or they won't get any logs.

Then we start with NewProductionConfig(), but we tweak it for a CLI,
making it easier to read.

Since we use a lot of DEBUG logs and these are capped by default
(default is INFO=0), then force the zapConfig.Level to be debug and
use a default filter that sticks with WARN and more critical errors
only.

This allows -l "debug+:*" or "*:*" to log all messages without the
need to create a config entry.