feat(tf): update convert functions and create/read operations
The convertion function was handling nested objects and arrays, but our
attributes don't handle those elements yet, this PR comments those
elements to allow the basic operations to execute.

The convertion function was also updated to handle special cases from TF
where the elements can be null when optional and not computed. And
handle cases where the state attribute is unknown due to never being
set.

Also updates the create/read operations to use the converter from and to
TF values.