fix(http): fix UnwrapResponse()
After the refactor this method was hard to use and thus was being used
incorrectly.

Clarify that the given data is actually a pointer to it, in order to
control the decoder with tagged structures.

Allocate a generic `any` dataPtr if none was given.

However, returning the data was leading to confusion, since it was a
pointer and not the final value. Then we use reflection to get the
final value out of it.