feat(ft): use cli auth credentials for tf
If we specify the access or refresh token into the terraform file the
access token will be valid until its expiration and the refresh will be
used only once but won't be updated since we always fetch the refresh
token from the tf file, and in every refresh we invalidate the previous
one.

If we don't specify neither of them the SDK will look for the
credentials file, that means, we can authenticate with the CLI and the
access token will already be provided to us. In case of expiration we
need to authenticate again with the CLI, but this is easier than setting
an env var for every usage.