refactor(tf): use new types to avoid mistakes
declare tfName and mgcName types and use them explicitly in maps, this
way we can avoid mistakes using terraform names in the mgc maps, and
vice versa.

Looking at a single line it should be possible to understand which one
to use, at least with variable names.

However, have the compiler to warn us as well.