fix(tf): allow computed only values to be updatable
This is necessary since after an update one of the computed only values
might change. If we set useStateForUnknown in all of them TF will
validate the Update Plan with the resulting State and an attribute was
updated but in the Update Plan it wasn't set as Unknown it will throw an
error.