fix(core): TerminatorExecutor shouldn't know about retries/interval
This was causing some confusion about it's concept and usage.

This executor is expected to wait until a termination state is
reached, such as in asynchronous operations that enter a transition
state before the final state is reached.

For some systems where a pub-sub may be used, there is no reason to
have "interval" or "maximum number of retries", it will only subscribe
to some event and wait the state to settle.

However in OpenAPI/REST, we're bound to call the operation multiple
times until it reaches a final state and there (and only in there) we
need these parameters.