refactor(core): Executor now returns a Result interface
BREAKING CHANGE!

The newly introduced Result interface allows implementations to
provide internal/extra details in addition to the actual value, such
as the schema that was used.

Default Formatting and Output options were also moved to Result
instead of the Executor, but we keep the Executor Wrappers, however
they now will call the original Executor's Execute() and then wrap the
Result instead.