chore: transofrm core.Parameters and Configs into simple typealiases
Debug sessions were bound to appear in the future due to validation
fails because of this new type. A TODO was added to consider patching
the lib that validates our data types to handle cases where a type is of
the correct structure but has a different signature due to type
assingments. For now, a typealias is used. In the future, when the
library does handle this case, it should be as simple as swaping

type Parameters = map[string]Value

to

type Parameters map[string]Value