feat(links): add support for chained Link commands
These commands can be called using the "!" operator. For example:

./mgc virtual-machine instances get --id some-id ! update --status shutoff

This will execute the 'get' action and then execute the 'update' link.
The link inherits some parameters from previous actions, so only the
'status' flag needs to be set in it, 'id' is already known. An arbitrary
number of links can be called.

Note that the subsequential links are links taken from the previous
executed action (be that a link or an executor), and NOT all from the
root action. So with 'action ! link1 ! link2', 'link2' will be searched
for in 'link1's links.