fix(core): fix and rewrite http.UnwrapResponse (continuation)
This function was confusing, as it was returning its values in two
fronts, via 'dataPtr' and via result return. this made it confusing to
work with.

Its implementation was also wrong, because it tried decoding a
non-pointer value, which would cause panics when accessing the reflect
value.

The rewrite sticks to only returning the result via normal return. If
the response data isn't convertible to T, it fails. To ensure that it
always works, regardless of the type of the result, it can be called via
'UnwrapResponse[any]'