feat(xml): add custom implementation of XML decoding
This interface uses the Golang XML decoder under the hood, but it
introduces a 'DisallowUnknownFields' method, just like the JSON decoder.
This will make it so that structs that attempt to decode an XML document
but lack all elements and attributes will fail.

It also prevents any decoding to native types such as 'int' or 'bool'.
This should, according to the docs, be the case in the official XML lib,
but errors are not returned when they should be