refactor(tf): add tfSchemaHandler interface and implement MgcResource
This interface is a general way of generating a TF Schema from the
'attributes' we provide. The interface that implements it must still
read its own input and output attributes, but the generation of the TF
Schema is handled elsewhere.

The 'schema.go' file was also cleaned up, all 'MgcResource'-specific
function implementations of the interface were moved to the
'resource.go' file, leaving 'schema.go' generic