feat(sdk): add blueprint infrastructure
Blueprints are YAML files that provides a tree of grouper and
executors.

Executors describe their schemas (parameters, configs and result), how
to create the final result and the execution steps.

Steps use other executors existing in the SDK (which can be Static,
OpenAPI or even other Blueprint), can be skipped, termination can be
awaited, it can be retried until a certain condition is met and it can
map parameters and configs from other steps, as well as the
blueprint's own data.