refactor(core): add DescriptorSpec, API BREAK!
This API break is long due as passing explicit parameters for name,
version and description (all strings) was cumbersome and error prone.

Add DescriptorSpec that groups those and also present a validation
method. It also comes with JSON tags to help serialization.

With that we add SimpleDescriptor, that takes a DescriptorSpec and
return its fields, this removes tons of useless code.