fix(config): don't instantiate new value for config get unmarshal
We used to instantiate a new value to unmarshal to in our config get
hook. This means that, when getting a partial config into a struct with
pre-defined values, those values would be lost EVEN IF the unmarshaled
data didn't have them, they would be overwritten to zero values.

This fix unmarshals the data directly into the object passed to
config.Get