fix: set proxy using DefaultTransport
Using the environment variable to set up the proxy would still not
be enough to make requests pass through it. This was found out to
be related to the previous transport overwriting the configuration
with a zero value. This has been changed by creating the transport
using http.DefaultTransport instead, with its ProxyFromEnvironment
call.