feat(core/utils): add flexible unordered slice compare
It's done on top of reflect.Value instead of any or generics so we can
more easily use it with unknown slices.

The generic version returns an error instead of a boolean so we can
more easily know what's happening.

The high-level version offers a bool return, so it's easier to use if
we just want to check without showing anything.