feat(scripts/transformers): add ConvertVersionTransformer
This handles conversion of 3.1 specification back to 3.0

This will get any `anyOf` or `oneOf` that contains a `{"type": "null"}`
entry and will remove it, promoting it to the parent as `nullable: true`
property.

Properties exclusiveMinimum and exclusiveMaximum, with actual numbers
will be converted to boolean and their value enforced into
minimum/maximum properties.

Examples array is converted to a single element.

This makes it pass JSON schema validation since it doesn't specify the
type "null".