feat(tf): allow resource operations without 'read' links to still apply state after
If a Resource operation was performed but it doesn't have a 'read' link,
the state appliance used to fail. Now, it can go on via the standard
Read call for the Resource. For this to work, it is crucial that the
original operation have all of is parameters and result applied first.
This is the case in the 'applyStateAfter' function