feat(cli): rewrite anyFlagValue as schemaFlagValue
This rewrite moves the schema and all the specific logic into this new
interface with multiple implementations.

The parser is smarter since it's based on the flag schema, so '1' will
parse as a string instead of an int, which would fail validation
later.

Objects and Arrays have accumulative parsers, and can deal with
CSV-like elements (separated by ',' or ';', in some cases such as
numbers/booleans, it can be separated by spaces as well).

Objects can be specified as `key:value` or `key=value`, which can be
surrounded by spaces.

Closes: #113, #114