chore(cli): remove todo about NoOptDefVal
I wrote this TODO when I copied that code from cobra, but now I
understand it better that if `NoOptDefVal` is not empty, then that
value will be used in place of the provided value.

Examples: given a `flag` with `DefValue=defval`, `NoOptDefVal=nooptdefval`:
 - --flag=v   -> v
 - --flag=    -> nooptdefval
 - (no flag)  -> defvalue

The common usage is with booleans, where `--x` implies in `--x=true`.

Closes: #307