fix: fix missing schema types
Schemas without a "type" was being misrepresented by the cli, fix that
by showing anyValue, anyOf, oneOf.

However, some cases were due broken schemas. While parameters can have
a parameter, that is optional and we should default to string. Do this
in a transformer, so our golang code is simpler