fix(cli): fix loading of flags, avoid ("no value")
When loading multiple flags and the LAST one returned a 'NoValue' error
(something expected by the flag loading code and HANDLED - it shouldn't
be returned to the user), the error would be mistakenly returned to the
User due to named returns. It would be set in the 'err' variable and
would never be unset. This commit fixes that