feat(core/http): add MGC_SDK_LOG_HTTP_PAYLOAD
With this we can log the HTTP request and response body. The wrappers
are a `progressive` one (logs as you read, safe as it won't blow the
memory, but it's cumbersome to use for JSON payloads) or `final` that
accumulates all data and logs at the end.

Since these may have significant impact in memory and behavior (if the
wrapper is incorrect), then do not add them by default. One must use
the $MGC_SDK_LOG_HTTP_PAYLOAD flag to control the behavior.