feat(core/schema): add ForEachObjectProperty()/CollectObjectPropertySchemas()
These should be the way we iterate over object properties so we don't
miss anyOf/oneOf handling anymore.

It will also sort the keys, making it easier to process.

The CollectObjectPropertySchemas() is a helper to find out all
possible schemas for a given property name. With that we can decide
what to do based on `Field` and `Index`.