feat(cli): explode object flags into children flags (proxy)
If a flag is an object with defined properties (not any map), then
create "children flags" that will just proxy their values to the
actual object.

This is recursive, so
`--network.vpc.name=X` proxies to
`--network.vpc=name=X` that will proxy to
`--network=vpc=name=X`.

In order to support such proxy the parser was changed to allow
`<root>=<child>=...=<leaf>=value`.

Alternatively the parser will also handle so called "path operations",
such as `--network=vpc.name=X`.

The examples will give precedence to these children flags, as more of
them seems to exist. The top-level flag example is only visible if
there is no child visible, otherwise it duplicates the value in
different forms, which is confusing.