fix(links): access and read linked command flags properly
Before this patch, an "unknown flag" error could occur when executing
linked commands due to root not knowing the specific flags of the
sub-command. We fix this error by setting the correct command for the
flag parser, at the right index.

The reason for this is because we were attempting to get those flags by
using c.root, but by switching to c.resolvedCmd, we can now find them
starting with index 1.

Since they start on index 1 and index 0 is the name of the command, we
updated the call to ParseFlags() so that it actually handles only flags,
as the function was designed for.