feat(tf): add support for property setters
Setters can be defined via links. The link name must start with "update"
and be followed by args separated by "/". There are 3 types of possible
setters:

- Simple
  Links to an Executor that will update the property

  update/<prop-name>

  Only one may be specified

- Enum
  Specifies different Executors, depending on the value being set in the
  property

  update/<prop-name>/<possibility-1>
  update/<prop-name>/<possibility-2>

  Many Enum possibilities may be specified

- Transition
  Specified different Executors, depending on the value being set AND
  the current value in the property.

  update/<prop-name>/<current-val-possibility-1>/<target-val-possibility-1>
  update/<prop-name>/<current-val-possibility-2>/<target-val-possibility-2>

  Many Transitions may be specified

Note that only one kind of setter may be specified for each property.
For example, if a property has a Simple property setter, it may NOT have
a Transition Setter as well