fix(tf): remove resource prefix from tf attrinute name
There are cases with two variables with different names refering to the same id.
For example: backup_id and id refering to the backups id. This would cause TF to
show two attributes for the same id, which is confusing and wrong.

This commit fixes this by removing the resource name from the input parameter,
in case there is an attribute with the same name without the preix in the response.
This means backup_id will be renamed to id

Issue #654