refactor(tf): implement standard way to run operations
An operation is now called MgcResourceOperation (interface). Each CRUD
operation can implement the interface in a different way. They must
have:

- CollectParameters
- CollectConfigs
- ShouldRun (should this run? if not, chain operations will still run)
- Run (execute and return result)
- PostRun (if Ran successfully, it's called. Generally applies state. If
  returns false or errors, chain operations will not be called)
- ChainOperations (if there are subsequential ops, return them here)

Arbitrary MgcResourceOperations can be create, it's not limited to the
CRUD ops. For example, the property setters are not an operation that
implements this interface and chains the next ones.