refactor(tf): leverage operation infra to "read after"
Every operation that wants to chain a read after it performs something
(generally through a Link, but sometimes through the normal root read
operation) can now use the 'MgcResourceReadAfter' operation, which will
perform the read and chain further operations after (with the read
result). The 'MgcOperationRunner' doesn't force a read anymore, any
operation that wants to read after may do it explicitly.

The 'applyStateAfter' function doesn't perform the read either, it just
applies the request parameters and result to the state