feat(progress): add 'UnitsReporter' infra
This new structure abstracts the need to keep references to channels in
the places where reporting progress in an async context is needed. The
only thing needed is to keep a reference to this reporter, call 'Start'
and defer 'End' and, when needing to report something, call 'Report'.
That's it, no need for go routines (already abstracted), channels, or
anything like that

This UnitsReporter supports updating the total value dynamically, and it
reports arbitrary units instead of bytes