feat(cli): improve auto-table output for nested columns
columns structure has a new `Parents` slice of strings. If that
exists, then more header lines are created with such parents and these
extra lines will have auto-merge enabled. That is, if a map entry
"FIELD" contains a subtable with 2 columns ("ID" and "NAME"), then it will
produce 2 header lines:

   | FIELD | FIELD | OTHER_FIELD
   | ID    | NAME  |
   | v1    | v2    | v3

With auto-merge, visually becomes:

   |     FIELD     | OTHER_FIELD
   | ID    | NAME  |
   | v1    | v2    | v3

For vertical tables (used to represent a single object), we do similar
logic but we merge horizontally as well:

   | FIELD | ID   | v1
   |       | NAME | v2
   | OTHER_FIELD  | v3

The automatic setup of `Parents` is done by creating groups of columns
inside columnsFromMap(), then we walk the groups and if they have more
than one column or the name is different from the map entry key, then
`Parents` is appended. This should allow multiple nesting levels to be
properly represented.

Closes: #839