fix(cli): fix object-storage objects list formatting
We should only promote slices if they are the only field of an
object, otherwise keep it as sub-table.

While promoting, if the column has no name, inherit the field name
instead.

Last, but not least simplify the buildVertically and make it more
uniform, we shouldn't check for `len(arr) == 1` here otherwise the
layout would change depending on how many items the server returns. We
always build vertically if the field value is NOT an array.